"feat: comentários para documentação javadoc"

+++ src/main/java/br/ufrn/imd/modelo/Board.java
@@ -2,46 +2,86 @@ package br.ufrn.imd.modelo;
 
 import java.util.ArrayList;
 
+/**
+ * A classe Board representa o tabuleiro de jogo para o jogo de batalha naval.
+ *  @version 1.0
+ */
 public class Board {
-    private ArrayList<Navio> navios = new ArrayList<>();
-    private ArrayList<Celula> celulas = new ArrayList<>();
+    private ArrayList<Navio> navios = new ArrayList<>(); // A lista de navios no tabuleiro
+    private ArrayList<Celula> celulas = new ArrayList<>();  // A lista de células no tabuleiro
 
+    /**
+     * O construtor da classe Board.
+     * Inicializa as células no tabuleiro.
+     */
     public Board() {
-        for (Integer x = 0; x < 10; x++) {
-            for (Integer y = 0; y < 10; y++) {
+        for (Integer x = (Integer) 0; x < 10; x++) {
+            for (int y = 0; y < 10; y++) {
                 this.celulas.add(new Celula(x, y));
             }
         }
     }
 
+    /**
+     * Obtém a lista de navios no tabuleiro.
+     * @return A lista de navios no tabuleiro
+     */
     public ArrayList<Navio> getNavios() {
         return navios;
     }
 
+    /**
+     * Define a lista de navios no tabuleiro.
+     * @param navios A lista de navios a ser definida no tabuleiro
+     */
     public void setNavios(ArrayList<Navio> navios) {
         this.navios = navios;
     }
 
+    /**
+     * Obtém a lista de células no tabuleiro.
+     * @return A lista de células no tabuleiro
+     */
     public ArrayList<Celula> getCelulas() {
         return celulas;
     }
 
+    /**
+     * Define a lista de células no tabuleiro.
+     * @param celulas A lista de células a ser definida no tabuleiro
+     */
     public void setCelulas(ArrayList<Celula> celulas) {
         this.celulas = celulas;
     }
 
+    // Métodos privados para gerenciar a posição dos navios e marcar células
+
+    /**
+     * Limpa as posições ocupadas por um navio no tabuleiro.
+     * @param navio O navio a ser removido do tabuleiro
+     */
     private void limparPosicoes(Navio navio) {
         for (Celula celula : navio.getCelulas()) {
             celula.setTipo(Tipo.AGUA);
         }
     }
 
+    /**
+     * Marca as posições ocupadas por um navio no tabuleiro.
+     * @param navio O navio a ser marcado no tabuleiro
+     */
     private void marcarPosicooes(Navio navio) {
         for (Celula celula : navio.getCelulas()) {
             celula.setTipo(Tipo.NAVIO);
         }
     }
 
+    /**
+     * Obtém a célula no tabuleiro com as coordenadas especificadas.
+     * @param x A coordenada x da célula
+     * @param y A coordenada y da célula
+     * @return A célula com as coordenadas especificadas, ou null se não encontrada
+     */
     private Celula getCelula(Integer x, Integer y) {
         for (Celula celula : celulas) {
             if (celula.getX().equals(x) && celula.getY().equals(y)) {
@@ -51,6 +91,14 @@ public class Board {
         return null;
     }
 
+    /**
+     * Posiciona um navio no tabuleiro com base nas coordenadas, navio e direção fornecidos.
+     * @param x A coordenada x
+     * @param y A coordenada y
+     * @param navio O navio a ser posicionado
+     * @param direcao A direção do navio
+     * @return true se o navio for posicionado com sucesso, false caso contrário
+     */
     public boolean posicionarNavio(Integer x, Integer y, Navio navio, Direcao direcao) {
         ArrayList<Celula> auxCelulas = new ArrayList<>();
         limparPosicoes(navio);
@@ -87,6 +135,12 @@ public class Board {
         return false;
     }
 
+    /**
+     * Marca uma célula no tabuleiro como atingida e verifica se todos os navios foram afundados.
+     * @param x A coordenada x da célula
+     * @param y A coordenada y da célula
+     * @return true se todos os navios foram afundados, false caso contrário
+     */
     public Boolean marcarCelula(Integer x, Integer y) {
         for (Celula celula : celulas) {
             if (celula.getX().equals(x) && celula.getY().equals(y)) {
